Bug Fixes
Fixed a few bugs: 1. related to imports, by adding import os, sys
sys.path.insert(0, os.path.abspath('.')) to the pySlope file, relative imports can be used which means the file also works with django or as a standalone file. 2. I set default values in models for color and material name and this seemed to fix the error with the material form. There was also another error were if a string is '' django actually interprets the value as None. So i changed the Material class to properly interpret a None value input. 3. Hide slope angle or hide slope length if not being used to define the slope.