Search data in html
Added functionality to have all searches displayed as a list of dictionaries (or really a list of objects)  which is saved as javascript and can be accessed with javascript. I also experimented and looks like it will be possible to add lines to the graphs created with python plotly with javascript. Also added a spinner icon and functionality to prevent multiple concurrent submits, and to let the user know the page is processing.